AI_model/requirements.txt
# New requirements.txt

# Core scientific and image processing
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
Pillow==10.4.0

# Deep Learning Framework
torch==2.4.1
torchvision==0.19.1
facenet-pytorch==2.5.3
mtcnn==1.0.0

# Face recognition utilities (optional but useful)
scikit-learn==1.5.2
scipy==1.13.1

# For backend communication
requests==2.32.3

# Progress bars and CLI utilities
tqdm==4.66.5

# Optional: For blinking/liveness detection via landmarks
# Original version (0.10.11) was incompatible with Python 3.12.
# Removing the version pin lets pip install the latest compatible version (e.g., 0.10.14 or later).
mediapipe

# for cnn
tensorflow>=2.12.0
scikit-learn>=1.2.0
matplotlib>=3.7.0

# facenet
#facenet-pytorch>=2.5.0
#torch>=2.0.0
#opencv-python>=4.8.0
#numpy>=1.24.0
#scikit-learn>=1.2.0
#pandas>=2.0.0

## faceArch
# insightface==0.7.3
# onnxruntime==1.23.2

